requirements.txt
# Core Django requirements
Django>=3.0,<4.0
djangorestframework
django-environ==0.8.1
django-allauth==0.51.0
django-sslserver
django-q
django-storages
django-unicorn
django-debug-toolbar
django-webpack-loader
django_pandas
dj-database-url==0.5.0

# Database
psycopg2>=2.8

# Google API packages
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client

# Python utilities
python-docx
openpyxl
asgiref==3.4.1
autopep8==1.6.0
gunicorn==20.1.0
pycodestyle==2.8.0
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
Unipath==1.1
whitenoise==5.3.0

# AI/ML packages
pinecone-client
langchain


# react 

django-webpack-loader==1.8.1

# Faker

Faker==19.13.0

